requirements.txt
opencv-python==4.11.0.86
numpy==2.0.2
pillow==11.1.0
emoji==2.14.1